// packages/apps-vtex/src/utils/types.ts
export interface Segment {
	campaigns: unknown | null;
	/** @description 1,2,3 etc */
	channel: string;
	priceTables: string | null;
	regionId: string | null;
	utm_campaign: string | null;
	utm_source: string | null;
	utm_medium: string | null;
	utmi_campaign: string | null;
	utmi_page: string | null;
	utmi_part: string | null;
	/** @description BRL, USD stc */
	currencyCode: string;
	/** @description R$, $ etc */
	currencySymbol: string;
	/** @description BRA, USA etc */
	countryCode: string;
	/** @description pt-BR, en-US etc */
	cultureInfo: string;
	channelPrivacy: "public" | "private";
}